Rearranged multiple tool buttons, part 2
