+++ src/components/LayerControlPanel/index.tsx
@@ -22,34 +22,6 @@
  * SOFTWARE.
  */
 
-import { CSSProperties, PropsWithChildren } from "react";
-import { SxProps } from "@mui/material";
-import Box from "@mui/material/Box";
+import LayerControlPanel from "./LayerControlPanel";
 
-const CONTAINER_STYLE: CSSProperties = {
-  position: "absolute",
-  display: "flex",
-  flexDirection: "column",
-  zIndex: 1000,
-};
-
-interface MapButtonGroupProps {
-  style?: CSSProperties;
-  sx?: SxProps;
-}
-
-export default function MapButtonGroup({
-  style,
-  sx,
-  children,
-}: PropsWithChildren<MapButtonGroupProps>) {
-  return (
-    <Box
-      className="ol-unselectable ol-control"
-      sx={sx}
-      style={style ? { ...CONTAINER_STYLE, ...style } : CONTAINER_STYLE}
-    >
-      {children}
-    </Box>
-  );
-}
+export default LayerControlPanel;
